deps: upgrade drizzle-orm and drizzle-kit to 1.0.0-rc.4
rc.4 is the current official RC (the rc dist-tag; v1.0.0-rc.4 release).
Adapting to its breaking changes:

- RQBv1 is removed from sqlite, so drizzle() no longer accepts the schema
  module. Nothing here uses db.query.*, so drop it from the six drizzle()
  calls (app client, unit and e2e setup, three scripts) and the schema
  generic from DB, TestDB and ImportDb.
- A bare blob() now defaults to JSON mode (drizzle#1064): reads parse and
  writes stringify, so passkey public keys and embeddings came back empty.
  All four byte columns get mode: 'buffer'. That changes no SQL:
  db:generate reports no schema changes, and existing rows (raw bytes
  written under rc.3) read back as before. The now-typed Buffer columns
  make two 'as Uint8Array' / 'as MemoryWithEmbedding[]' assertions
  redundant, so they're gone.
- sqlite errors are wrapped in DrizzleQueryError; handled in the previous
  commit.

CLAUDE.md records the blob-mode and error-wrapping traps plus the
hand-upgrade policy. The unit suite passes (298 files), as do lint and
pnpm check.

+++ src/lib/server/db/queries/mcp-credentials.ts
@@ -25,7 +25,7 @@ export function getMcpCredential(userId: string, serverId: string): string | nul
 		.get();
 	if (!row) return null;
 	try {
-		return decryptSecret(row.ct as Uint8Array);
+		return decryptSecret(row.ct);
 	} catch (e) {
 		// Key rotated, or the row predates a key change / is corrupt. Treat as
 		// "no usable credential" so the user is prompted to re-enter rather than
